Fix google login issues, add login check to tests
Google login did not use GoogleConfirmAuth fragment. A check for its
presence was added into LoginTest. The code of GoogleConfirmAuth was
simplified.

A check for logged in user was added to tests using login. This change
was added because tests failed on staging-new.slepemapy.cz/ site. The
reason is that there is user logged in by default.

Minor changes that fix some browser synchronization issues.